feat: refactor conversation types and update imports for clarity
